Licensing conflict resolved
In this release the DHTMLX library is removed, and replaced with the jQuery UI plugin. DHTMLX has a GLPv2 license.

Most of the functionality remains exactly the same. OpenWindow/Alert/Confirm/Prompt is replaced, so there no changes if you are re-using these in your own modifications. The DHTMLX popup feature cannot be replaced as-is, however a new PopupMenu() UI feature has been added to replace most of the previously used popup menus. Some design changes had to be made to accommodate some of the limitations.

https://github.com/InnovateAsterisk/Browser-Phone/issues/152